barraNavegacion
barras de navegación corregidas
