+++ frontend/src/views/layouts/auth/estiloLayout.css
@@ -18,9 +18,6 @@
     overflow: hidden;
   }
   
-  
-  
-  
   /* .barra-lateral {
     font-size: 16px;
   }
@@ -65,13 +62,3 @@
     max-width: 100%; 
     max-height: 100%;
   }
-  
-  
-  
-  
-  
-  
-  
-  
-  
-
