+++ frontend/src/views/layouts/unauthorized/unauthorized.layout.jsx
@@ -3,9 +3,9 @@ import { Routes, Route, Navigate } from "react-router-dom";
 import LoginPage from "../../pages/loginPage/loginPage";
 import RegisterPage from "../../pages/registerPage/registerPage";
 import MainPage from "../../pages/mainPage/MainPage";
-import PatientLayout from "../auth/patientLayout/patient.layout";
-import DoctorLayout from "../auth/doctorLayout/doctor.layout";
-import LaboratoryLayout from "../auth/laboratoryLayout/laboratory.layout";
+import PatientLayout from "../auth/patient.layout";
+import DoctorLayout from "../auth/doctor.layout";
+import LaboratoryLayout from "../auth/laboratory.layout";
 
 export default function UnauthorizedLayout() {
   return (
